can't update or delete a zikresource created by somebody else

+++ src/main/zikresource/zikresource-blo.ts
@@ -1,3 +1,4 @@
+import { AddedByUser } from '../user/user';
 import { ZikStockError } from '../zikstock-error/zikstock-error';
 import { Zikresource } from './zikresource';
 import { ZikresourceDAO } from './zikresource-dao';
@@ -34,10 +35,16 @@ export class ZikresourceBLO {
         return result;
     }
 
-    async deleteOneZikresource(id: string): Promise<void> {
+    async deleteOneZikresource(id: string, userWhoWantsTheDeletion: any): Promise<void> {
         let zikresource = await this.getOneZikresourceById(id);
         if (zikresource == null) {
             throw new ZikStockError("404-1");
+        }
+        let isCreatedBySameUser = userWhoWantsTheDeletion.addedBy != null 
+            && userWhoWantsTheDeletion.addedBy.email != null
+            && await this.hasSameOwner(id, userWhoWantsTheDeletion.addedBy?.email);
+        if (!isCreatedBySameUser) {
+            throw new ZikStockError("400-5");
         }
         let deleted = await this.zikResourceDAO.delete(zikresource);
         if (!deleted) {
@@ -48,6 +55,11 @@ export class ZikresourceBLO {
     async updateOneZikresource(id: string, data: any): Promise<Zikresource | undefined> {
         // Prerequisites:
         this.checkIfDataAreValid(data);
+        let isCreatedBySameUser = data.addedBy != null 
+            && data.addedBy.email != null && await this.hasSameOwner(id, data.addedBy?.email);
+        if (!isCreatedBySameUser) {
+            throw new ZikStockError("400-5");
+        }
         // if prerequisites are ok, we don't have exception, so we do it:
         let zikResourceUpdated = await this.zikResourceDAO.updateOne(id, this.buildZikresourceInstance(data));
         return zikResourceUpdated;
@@ -69,7 +81,15 @@ export class ZikresourceBLO {
         if (data.type) { zikresource.type = data.type; }
         if (data.artist) { zikresource.artist = data.artist; }
         if (data.tags) { zikresource.tags = data.tags; }
+        if (data._id && data.email && data.displayName) {
+            zikresource.addedBy =  new AddedByUser(data._id, data.email, data.displayName);
+        }
         return zikresource;
+    }
+
+    private async hasSameOwner(zikResourceId: string, userEmail: string): Promise<boolean> {
+        const zikResource = await this.getOneZikresourceById(zikResourceId);
+        return (zikResource != null && zikResource.addedBy?.email === userEmail);
     }
 
 }
